ref: Move strategies out of streaming module
The streaming module is kind of redundant and just makes imports more verbose - this whole library is
about streaming anyway and the batching strategies will be getting deprecated. All the strategies
will now live directly under the strategies module.
