+++ arroyo/processing/strategies/__init__.py
@@ -3,5 +3,17 @@ from arroyo.processing.strategies.abstract import (
     ProcessingStrategy,
     ProcessingStrategyFactory,
 )
+from arroyo.processing.strategies.collect import CollectStep, ParallelCollectStep
+from arroyo.processing.strategies.filter import FilterStep
+from arroyo.processing.strategies.transform import ParallelTransformStep, TransformStep
 
-__all__ = ["MessageRejected", "ProcessingStrategy", "ProcessingStrategyFactory"]
+__all__ = [
+    "CollectStep",
+    "ParallelCollectStep",
+    "FilterStep",
+    "TransformStep",
+    "ParallelTransformStep",
+    "MessageRejected",
+    "ProcessingStrategy",
+    "ProcessingStrategyFactory",
+]
