+++ arroyo/processing/strategies/collect.py
